refactor: clean types/equipment barrel, extract IEquipmentItem and aggregation logic
